Added data downsampler for plotting
Added a data downsampler class (DataDownSampler) in resample.py for
plotting less points in the window for speed.  When the window is
update, it will redraw keeping the number of points constant unless you
are zoomed in too far where downsampling is not required.

Made necessary changes to data_plotter.py and app.py to accomodate this
new functionality.

Known Issues:
	If you are zoomed in already and then check Share X-axis, the
	next axis update will be weird for windows that are already
	zoomed in.  For now, I recommend resetting to home before
	clicking Share X-axis.