Added a QColorDialog to edit configuration functionality
Added a new row to the EditConfigDialog in edit_configuration.py for
picking colors for each element in a plot.  This configuration data is
now part of the config file.

If color is not listed for a signal in the config file, it will be given
one of the 10 default colors automatically.  This will lead to repeat
colors if you have more than 10 signals in a given plot.