Better no data handling, UNTESTED
First attempt at better data handling, especially with no data.

This is untested because I can't log into skywalker at the moment.

The methodology is to return None if there is no data.  The None's will
get stored into the lists / corresponding entry in the self.data
dictionary.  I wrote a helper function in mdsplus_helpers.py to check
the data dictionary for actual data.  If there is no actual data,
self.data will be written to None and no plotting will happen.

Ideally, if the user makes a change externally and then presses update,
the app should ask for data again.  This is a change from before.
Hopefully, this will work seamlessly.