Added grid spec with uneven rows
This is still a work in progress.  I probably have added a bunch of bugs
that I have not found yet.

The config file now has a col key which is a list of 6 numbers
corresponding to the number of rows in that column.  The first column
must be 1 or higher. The rest could be zero.

I also made the edit configuration window a little easier to read.  It
now specifically says Col: # Row: # to make it obvious which subplot is
being configured.

Added a helpers.py file to do a least common multiple calculation for a
list of numbers.