Tracking down a locking bug
Ethan and I have witnessed a bug in the control room where the gui locks
up while in share-x and auto update mode.  I assume the culprit is auto
update and not share-x.

I have added a QTimer to query the mdsplus event watching thread to ask
if it is still alive when auto update is active.  I have yet to
replicate the bug at home and the thread has always been active.

In response to that knowledge, I added some extra logic inside in
app.py.  When it is time to spawn all of the mdsplus calls, it will call
a brand new function in mdsplus_helpers.py called check_open_tree.
Before any data calls go out, it must get a True back saying the tree is
able to be opened.  If it doesn't, then it will gracefully move on and
say there was an error opening the shot and setting self.data to None.

I'm hoping this fixes some of the issues we have been having.

Unfortunately, all of these fixes are inside this branch that might have
bugs of its own.  Good Luck!