Moved config parsing to config module
All of the config parsing that was in app.py is now in config/parser.py

Did a little more code clean up.  Moved the figure creation to
data_plotter.py.

Tried to fix some of the logic for the acquiring_data flag.

Added some comment strings to explain logic.  Started adding
docstrings to app.py.