Bug fix with xlim and ylim enable/disable
The way I check to see if xlim and ylim should be enabled is to see if
the key is in the config dictionary.  I wasn't deleting the key and
value from the dictionary when it should be.

Note on removing keys from dictionary:

dictionary_variable.pop(key_to_remove, None) will remove key_to_remove
and return the value of dictionary_variable[key_to_remove].  If
key_to_remove is not in dictionary_variable, then it will return None.

I changed the deleting code for signals to this intead of using python's
del function.