Add empty_lru_cache function
Added a function called empty_lru_cache function to empty the cache for
the decorated _retrieve_signal function.  I want the user to be able to
regrab data for the same shot if a modification has been changed on the
tree side.

Because data.Data is now truthy, I changed the check_data_dictionary to
leverage that functionality instead of checking against None.

Also, removed some old print statements.