Cast time and data as numpy ndarrays
I've decided to just leverage numpy for my type checking. I really only
want to deal with numpy arrays, so I removed the isistance check against
Iterable.

The truthyness now is simplier and just checks if len(self) is greater
than 1.  An implementation detail that I'm not sure if I like is it will
throw a ValueError if the lengths of time and data don't match.