Add concurrent.futures functionality
I'm replacing using the QThreadPool for handling all of the MDSplus
calls with a concurrent.futures.ThreadPoolExecutor.

There is a new function in mdsplushelpers called retrieve_all_data that
uses the ThreadPoolExecutor.  A keyword is a progress_signal. This is a
Qt signal that can be connected back to a progress bar to keep
functionality from before.  I should probably wrap the emitting in a
try except clause.

retrieve_all_data is called from an instance of a Worker using the
original QThreadPool in app.py. The done signal is connected to
handle_mdsplus_data.

I've removed handle_returning_data because it is no longer needed.  The
progress bar is updated via the new function update_progress_bar which
is connected to the worker progress signal.

I modified workers.py to pass its progress signal to retrieve_all_data.

ToDo:
    Decide on an timeout implementation detail.  I'm thinking about
    adding it to a global configuration.  It is nice to have, but it can
    take a long time to grab all of the data for long cathode plasmas.